requirements/requirements_base.txt
django==4.2.7
django-debug-toolbar==4.2.0
python-dotenv==1.0.0
stripe==7.6.0
